// src/types/favorite.ts
import { LocationType } from './location';
import { CityType } from './city';

export type FavoriteType = {
  id: string;
  title: string;
  type: string;
  price: number;
  city: CityType;
  location: LocationType;
  isFavorite: boolean;
  isPremium: boolean;
  rating: number;
  previewImage: string;
};